Rotation and Shooting
Includes my fix to rotation along with an implementation of shooting
using a simple line trace. Once a swan car is shot 5 times, it will
explode.

Blueprint changes: turret_topBP, swanCarBP, and a new BP for GUI
crosshair

I also reorganized the unreal project and added the starter pack so we
can make use of the animations and sound effects.