Two Bug Fixes
Fixed turret detaching from blue car when it respawns

Fixed blast radius logic. It was set up so that it checks how far away a
hit was from an input actor. However, that check only happened when it
HIT the actor. To fix this, I checked blast radius before the casts
instead of after each cast. After it checks to see if the target actor
was a certain distance from the hit point, it tries to cast to the
target actor (which, effectively, figures out which kind of target it
is). Most hits are within 75cm of the center of the defaultCar.

Because of this second change, I raised the health of the
player-controlled car to 100.