Character Selection, AITurret, and Turret Animation
Got the character selection working somewhat using a GameInstance.
The
way that GameInstance works is that it persists even between
level
switches (where passed values are lost) so I pass the values there
and
once the level is loaded up I reference them inside of a
Blueprint.
Once it is fully implemented, I will only reference them in
the Level Blueprints since those
are the blueprints that take care of
possession. Maybe we can use 2
arrays of pawns (each containing each Car
and Sentry references
respectively) and every time one gets set we
delete it from the array,
thus leaving only the unassigned ones for the
AICars to choose from.

Now, we just need to get the AICars to be the
same as the
player-controlled one so that the player can just possess
whichever car
has their characters instead of doing it a much harder
way.

For AITurret, I removed Turret2 for the player-controlled car
since
Turret1 works just as well and will let us have cars that are
exactly
the same. Also, I converted the BaseTurret Actor Blueprint into
a Pawn
Blueprint so that it could be possessed.
I also added a muzzle flash using the MuzzleFlash component by
putting
that component on front of the barrel of the gun.

I think I did
some other stuff too but I can't remember at this point. I had
everything done and went to push, but there was one file that was too
big and it sort of deleted everything that I did. So, I had to redo it.
Luckily I remembered most of the changes that I made. I'll fix the rest
tomorrow.