Project Organization, Car Switching, and Build Changes
I reorganized the project and moved things around:

All of the textures and materials (that are not part of "Starter
Content" are now found in the "Textures" folder.

"AICar" is now its own top-level folder called "DefaultCar_AI" and
Waypoints have been migrated here as well.

All levels have been moved to the "Levels" folder.

Besides that, I got the car switching working by using "set location and
rotation" instead of "teleport". I'm not sure how we are easily going to
set the velocity of the car. We have to mess around with the
WheeledVehicleComponent values or something like that.

Also, I had to change a lot in the WheelsInTheSky level because the
level blueprint was not set up correctly. It should work now with no
problems (except for the turret doesn't have x rotation but who cares).

You _should_ be able to build now.