Impulse Forces on Car Switch
If the player switches from the turret to the wheel or vice versa, a
forward impulse force is added to the car in order to make up for the
fact that the car's velocity restarts at 0 due to the switch.

Things to do:
add a cooldown to the switching mechanic
make the shooting slower and more powerful.