Turret Fix, Level Blueprint Optimizations, UI Improvements
I was able to fix the turrets by switching yaw and roll along with a
couple other things. I was also able to get the player turret clamping
fixed by testing the new yRot with the current relativeRotation before
adding yRot to the relativeRotation. If it takes it over/under a limit,
I set yRot to 0 so that it does not go past that angle but can still
rotate on the x axis.

I changed the level blueprint quite heavily. The most notable change was
getting rid of the event dispatcher and using the game instance instead
to store all of the health values, scores, and positions in the race.
Positions have not been implemented yet but it will be a series of
trigger colliders along the map where the player increments their
integer position if they go through one. Then, this value will be
compared against the rest of the positions for a placing.

The usage of the game instance makes the gui stuff a lot easier. I was
able to get the player health consistent across the defaultCar and
playerAICar. With that, the player's health is on the screen as 'Red'
but will be whatever player that they are. Along with this, I keep the
healths of all of the AICars so that if we wanted to we could have
health bars on/below/above the other cars as they are driving as an
indicator of their health.