Main Menu Update, Loading Screen, Health Fix, Driving Optimizations
Implemented the Storybook main menu with an animated flipping page.

Added a loading screen to the main menu.

Fixed the player health so that it updates with its AI Car.

Fixed driving so it sucks a little less I guess. Less "driving on ice"
more "it's a tad hard to turn now"