Tweaked Vehicle Movement
Set the defaultCar to have a max RPM of 9500 to the AICar's 10000. This
lets the AICars keep up with the defaultCar better.

Lowered the friction values for the wheels and lowered the mass of all
of the cars.

We may want to consider a single lap where the finish is not accessible
from the start for the maps. This is because placings are just not going
to happen in the next few days. The amount of code needed for the
triggers vastly outweighs the benefit of having it in the game (we would
need an event for every trigger and we also need to figure out which
actor interacted with the trigger).

Then, we can keep track of which player crossed the finish line (only
once) and print out the order after the race is over as our placings.

We might want to do a simple tutorial level introducing the driving
mechanics, shooting, and switching.

Primary focus should be updating waypoints for the player's AICar when
the player reaches one as the defaultCar.